requirements.txt
APScheduler==3.6.3
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
pip==22.3.1
python-dotenv==0.21.0
python-telegram-bot==13.15
pytz-deprecation-shim==0.1.0.post0
pytz==2022.7
requests==2.28.1
setuptools==65.6.3
six==1.16.0
tornado==6.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
wheel==0.38.4
